+++ packages/rma-frontend/src/app/sales-ui/add-sales-invoice/add-sales-invoice.page.spec.ts
@@ -8,7 +8,10 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { of } from 'rxjs';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { STORAGE_TOKEN } from '../../api/storage/storage.service';
+import {
+  StorageService,
+  STORAGE_TOKEN,
+} from '../../api/storage/storage.service';
 import { SalesService } from '../services/sales.service';
 import { ItemPriceService } from '../services/item-price.service';
 
@@ -63,6 +66,12 @@ describe('AddSalesInvoicePage', () => {
               getItem: (...args) => Promise.resolve('Item'),
               getItems: (...args) => Promise.resolve({}),
             }),
+          },
+        },
+        {
+          provide: StorageService,
+          useValue: {
+            getItem: (...args) => Promise.resolve('ITEM'),
           },
         },
         { provide: STORAGE_TOKEN, useValue: {} },