+++ packages/rma-server/src/system-settings/entities/server-settings/server-setting.dto.ts
@@ -190,4 +190,5 @@ export class ServerSettingsDto {
     type: 'object',
   })
   brand: BrandSettingsDTO;
+  backdatedInvoices: boolean;
 }