+++ packages/rma-warranty/src/app/warranty-ui/shared-warranty-modules/service-invoices/service-invoices.component.ts
@@ -23,6 +23,7 @@ export class ServiceInvoicesComponent implements OnInit {
   dataSource: ServiceInvoiceDataSource;
   permissionState = PERMISSION_STATE;
   total: number = 0;
+  disableRefresh: boolean = false;
   displayedColumns = [
     'invoice_no',
     'status',
@@ -52,7 +53,18 @@ export class ServiceInvoicesComponent implements OnInit {
   ngOnInit() {
     this.invoiceUuid = this.route.snapshot.params.uuid;
     this.dataSource = new ServiceInvoiceDataSource(this.serviceInvoice);
-    this.dataSource.loadItems(this.invoiceUuid);
+    this.dataSource.loadItems({ warrantyClaimUuid: this.invoiceUuid });
+    this.dataSource.disableRefresh.subscribe({
+      next: res => {
+        this.disableRefresh = res;
+      },
+    });
+  }
+
+  syncDocStatus() {
+    this.dataSource.syncDocStatus().subscribe({
+      next: res => {},
+    });
   }
 
   getUpdate(event) {
@@ -65,7 +77,7 @@ export class ServiceInvoicesComponent implements OnInit {
       }
     }
     this.dataSource.loadItems(
-      this.invoiceUuid,
+      { warrantyClaimUuid: this.invoiceUuid },
       sortQuery,
       event.pageIndex,
       event.pageSize,